Changed locks to ReentrantLock instead of readwrite lock

+++ src/SHoxy/Cache/CacheCleaner.java
@@ -1,13 +1,10 @@
 package SHoxy.Cache;
 
 import java.io.File;
-import java.nio.file.Files;
 import java.util.Collection;
 import java.util.Date;
 import java.util.Map;
 import java.util.Random;
-
-import java.util.concurrent.*;
 import java.util.concurrent.locks.Lock;
 
 public class CacheCleaner implements Runnable {
@@ -54,33 +51,30 @@ public class CacheCleaner implements Runnable {
     }
     
     private boolean isTooOld(CachedItem item) {
-        Lock readLock = item.getLock().readLock();
-        readLock.lock();
         boolean result = false;
-        try {
+
+        Lock itemLock = item.getLock();
+        if (itemLock.tryLock()) {
             Date now = new Date();
             Date lastRequested = item.lastTimeRequested;
             
             result = (now.getTime() - lastRequested.getTime() <= deleteTimerSeconds * milliseconds);
-        } catch (Exception e) {
-            result = false;
-        } finally {
-            readLock.unlock();
+            itemLock.unlock();
         }
+        
         return result;
     }
 
     private void removeFromCache(CachedItem item) {
         // Try to get a lock and remove the item from the cache.
         // If that fails, give up, because it's in use still
-        Lock writeLock = item.getLock().writeLock();
-        writeLock.lock();
-        try {
+        Lock itemLock = item.getLock();
+        if (itemLock.tryLock()) {
             String fileURL = CachedItem.parseURLToFileName(item.URL);
             deleteFile(fileURL);
             cache.remove(item.URL);
-        } finally {
-            writeLock.unlock();
+            
+            itemLock.unlock();
         }
 
     }
